+++ modules/iam-access-groups/src/main/java/com/ibm/cloud/platform_services/iam_access_groups/v2/model/AddMembershipMultipleGroupsResponseGroupsItem.java
@@ -32,7 +32,7 @@ public class AddMembershipMultipleGroupsResponseGroupsItem extends GenericModel 
   /**
    * Gets the accessGroupId.
    *
-   * The Access Group that the member is to be added to.
+   * The access group that the member is to be added to.
    *
    * @return the accessGroupId
    */