+++ modules/iam-access-groups/src/main/java/com/ibm/cloud/platform_services/iam_access_groups/v2/model/DeleteFromAllGroupsResponseGroupsItem.java
@@ -32,7 +32,7 @@ public class DeleteFromAllGroupsResponseGroupsItem extends GenericModel {
   /**
    * Gets the accessGroupId.
    *
-   * The Access Group that the member is to be deleted from.
+   * The access group that the member is to be deleted from.
    *
    * @return the accessGroupId
    */