+++ modules/iam-access-groups/src/main/java/com/ibm/cloud/platform_services/iam_access_groups/v2/model/GetAccessGroupOptions.java
@@ -115,7 +115,7 @@ public class GetAccessGroupOptions extends GenericModel {
   /**
    * Gets the accessGroupId.
    *
-   * The Access Group identifier.
+   * The access group identifier.
    *
    * @return the accessGroupId
    */
@@ -126,7 +126,9 @@ public class GetAccessGroupOptions extends GenericModel {
   /**
    * Gets the transactionId.
    *
-   * An optional transaction id for the request.
+   * An optional transaction ID can be passed to your request, which can be useful for tracking calls through multiple
+   * services by using one identifier. The header key must be set to Transaction-Id and the value is anything that you
+   * choose. If no transaction ID is passed in, then a random ID is generated.
    *
    * @return the transactionId
    */