+++ modules/iam-policy-management/src/test/java/com/ibm/cloud/platform_services/iam_policy_management/v1/model/PatchPolicyOptionsTest.java
@@ -11,32 +11,39 @@
  * specific language governing permissions and limitations under the License.
  */
 
-package com.ibm.cloud.platform_services.iam_access_groups.v2.model;
+package com.ibm.cloud.platform_services.iam_policy_management.v1.model;
 
-import com.ibm.cloud.platform_services.iam_access_groups.v2.model.Rule;
-import com.ibm.cloud.platform_services.iam_access_groups.v2.model.RuleConditions;
-import com.ibm.cloud.platform_services.iam_access_groups.v2.model.RulesList;
-import com.ibm.cloud.platform_services.iam_access_groups.v2.utils.TestUtilities;
+import com.ibm.cloud.platform_services.iam_policy_management.v1.model.PatchPolicyOptions;
+import com.ibm.cloud.platform_services.iam_policy_management.v1.utils.TestUtilities;
 import com.ibm.cloud.sdk.core.service.model.FileWithMetadata;
-import com.ibm.cloud.sdk.core.util.DateUtils;
 import java.io.InputStream;
-import java.util.ArrayList;
-import java.util.Arrays;
 import java.util.HashMap;
 import java.util.List;
 import org.testng.annotations.Test;
 import static org.testng.Assert.*;
 
 /**
- * Unit test class for the RulesList model.
+ * Unit test class for the PatchPolicyOptions model.
  */
-public class RulesListTest {
+public class PatchPolicyOptionsTest {
   final HashMap<String, InputStream> mockStreamMap = TestUtilities.createMockStreamMap();
   final List<FileWithMetadata> mockListFileWithMetadata = TestUtilities.creatMockListFileWithMetadata();
 
   @Test
-  public void testRulesList() throws Throwable {
-    RulesList rulesListModel = new RulesList();
-    assertNull(rulesListModel.getRules());
+  public void testPatchPolicyOptions() throws Throwable {
+    PatchPolicyOptions patchPolicyOptionsModel = new PatchPolicyOptions.Builder()
+      .policyId("testString")
+      .ifMatch("testString")
+      .state("testString")
+      .build();
+    assertEquals(patchPolicyOptionsModel.policyId(), "testString");
+    assertEquals(patchPolicyOptionsModel.ifMatch(), "testString");
+    assertEquals(patchPolicyOptionsModel.state(), "testString");
   }
+
+  @Test(expectedExceptions = IllegalArgumentException.class)
+  public void testPatchPolicyOptionsError() throws Throwable {
+    new PatchPolicyOptions.Builder().build();
+  }
+
 }