+++ modules/iam-policy-management/src/test/java/com/ibm/cloud/platform_services/iam_policy_management/v1/model/PolicyTest.java
@@ -45,5 +45,6 @@ public class PolicyTest {
     assertNull(policyModel.getSubjects());
     assertNull(policyModel.getRoles());
     assertNull(policyModel.getResources());
+    assertNull(policyModel.getState());
   }
 }